Refactor login to match new parent response (WIP)
the login and get profile response now return the list of children so no need to call get students home to get the array of children. So, the code is refactored to match the new flow.
the home activity only calls to refresh the firebase token and the array list of students is sent through the presenter to the activity